backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
spacy==3.7.2